requirement.txt
pandas
cx-Oracle
cryptography
XlsxWriter
PyYAML
pywin32; platform_system=="Windows"
